chore: reset the repo to an empty state
Remove all the previous code as the glium project has been deprecated,
Remove appveyor config.
Update copyright date.
Remove API docs and Coveralls badges from the README.
Update the README to reflect the updated project goals, and add a notice
regarding the rewrite.
Add a vision statement for the project. Closes #95.
Strip down the Travis CI configuration.

meaning that a complete rewrite of the rendering system is necessary.
As the game is not particularly far advanced at this point, rewriting the
rendering system means rewriting most of the existing code; so I have
decided to scrap all of the existing code and start from scratch.
Parts of the old code may be re-used in the future.