feat: add a basic game loop
Add time as a dependency.
Add a timer to calculate elapsed ticks between frames.
Add a basic game loop which is executed a specified number of times per
second.

Add SDL2 as a dependency with the "image" feature enabled.
Add sdl2 and sdl2_image Arch Linux packages to prerequisities list in
README.
Add code to initialize an SDL2 window of a specified size prior to starting the
game loop.
Add code to handle SDL2 quit events and specified keys which are used to quit
the game.