feat: add recording/playback and basic voxel chunks
Add text rendering via sdl2_ttf for the rendering of debug info to the
screen.
Load the included NotoSans-Regular.ttf font for rendering.

Use bincode to implement an input playback/recording system by recording
the game state and any subsequent inputs to the disk, and then loading
them back and repeating them in a loop.

Add camera-like motion by adding/subtracting an X & Y offset from the
rendered position of the chunk.

Add functions for loading individual voxels or rows of voxels from a
chunk.

Add a `WorldGenerator` struct to generate a world, represented currently
by a single chunk comprising a 32x32x32 voxel cube.
Add a height map which is generated by the world generator prior to
being used by the world generator to generate chunks.

Add basic input/controller state handling by introducing structs to
store the input states across frames, via the `GameInput`,
`ControllerInput`, and `GameButtonState` structs.
Extract event processing out of the main loop.

Add a `Chunk` struct to store a (semi-sparse) cube of voxel data.
Add a `render_chunk` function to render a chunk (in a very naive way).

Add a `State` struct to encapsulate all of the state relevant to the
current game execution.

Add a `Game` struct to encapsulate all of the state relevant to the game
itself.

Add a `VOXEL_ROWS` global variable containing an array of "packed" rows
of voxel data for each available material. This variable is used to
provide instances of a row of voxels where all of the voxels are of the
same type of material. This allows for massive memory savings as we no
longer need to store packed voxel rows in the `Chunk` struct, and can
instead fetch them from this array as necessary. This is sort of similar
in principle to run-length encoding.

Add constants for the commonly used colours white, black, and grey.

Add a `GLOBAL_RUNNING` global variable to track whether or not the game
loop has been terminated.

Add functions to draw rectangles and textures to the screen.

Add a `serde_array` macro found on github, allowing for the
serialization of arbitrarily large fixed size arrays with serde.